Width-match PostgreSQL integer columns instead of uniform BIGINT (#603) (#629)
* Width-match PostgreSQL integer columns instead of uniform BIGINT (#603)

schema.cppm's integer_type<D>() mapped every C++ integer-stored field
(short, char, int, unsigned int, int64_t, enums, chrono durations) to
BIGINT regardless of width, wasting 2x storage/wire bytes for narrow
fields and giving up DB-side overflow protection. PG integer columns
are now width-matched per C++ type (short/1-byte -> SMALLINT,
int/unsigned short -> INTEGER, int64_t/unsigned int/long long ->
BIGINT; enums and chrono::duration follow their underlying/rep type).
SQLite is unaffected (single dynamic INTEGER storage class already).

A DB-generated single-column integer PK stays hardcoded to BIGINT on
PG regardless of its declared C++ width (unchanged, pre-existing
behavior), so anything mirroring that PK's type (a junction column, a
composite-FK part column) must mirror the pinned width too, or PG's
FOREIGN KEY type check rejects the CREATE TABLE. Two new predicates
(has_pinned_bigint_pk, part_pinned_bigint) gate this, cross-checked by
a static_assert where the pin originates.

Breaking change to already-generated PostgreSQL schemas, accepted
because no production Storm database exists yet.

This container lacks the project's custom clang-p2996 toolchain, so
this change could not be locally built or tested — verified by manual
tracing, cross-referencing existing test expectations, and two rounds
of storm-code-reviewer/storm-sql-reviewer. CI is the first real build.


* Fix clang-format and clang-tidy findings in #603 (real toolchain verified)

Got Docker working in this session (docker/ci/Dockerfile + a local CA-trust
layer for the proxy) and pulled the project's actual clang-p2996 toolchain,
so the #603 change from the previous commit could finally be built and
tested for real rather than only manually traced.

Full build + full test suite (SQLite and live PostgreSQL, 3093 tests) is
green - 0 failures, 100% coverage. Two things the real toolchain caught
that manual review couldn't:
- clang-format: one line I'd hand-formatted didn't match the project's
  pinned clang-format version.
- clang-tidy bugprone-branch-clone: integer_type<D,W>/integer_type_nn<D,W>'s
  if-constexpr chain had two branches that coincidentally both return
  "INTEGER" (SQLite, and PG's Regular width) - not a real clone, but the
  same false-positive shape this file already hit once before (see the
  .clang-tidy comment on the FK dialect dispatch). Refactored to the same
  fix used there: a constexpr array indexed by [width][dialect] instead of
  an if-constexpr chain, so there's no branch for the checker to compare.


* Fix SonarCloud S7035: use std::to_underlying instead of static_cast<int>

PR #629's SonarCloud analysis flagged the two IntWidth->array-index casts
in integer_type<D,W>/integer_type_nn<D,W> (added earlier in #603) against
cpp:S7035 (use std::to_underlying to cast enums to their underlying type,
not static_cast). Both were plain enum-to-int casts for array indexing;
std::to_underlying is the correct, self-documenting way to spell that.


---------

+++ tests/schema/test_composite_primary_key.cpp
@@ -267,7 +267,8 @@ namespace {
 // and NO column-level PRIMARY KEY — reusing the junction-table DDL pattern. The PK
 // part columns are ordinary NOT NULL columns.
 //
-// The two dialects differ only in the int column type (SQLite INTEGER, PG BIGINT) —
+// Both dialects render "INTEGER" here: OrderItem's parts are plain `int`, and PG's
+// width-matched integer mapping (#603) picks INTEGER for a 4-byte signed source —
 // the ordinary per-dialect mapping, unchanged by the composite path. Neither carries
 // the SQLite AUTOINCREMENT nor the PG GENERATED AS IDENTITY decoration: every part of
 // a composite key is caller-supplied, so no part can be DB-generated.
@@ -294,7 +295,7 @@ TEST(CompositePrimaryKeyDDL, SqliteEmitsTableLevelPrimaryKey) {
 
 TEST(CompositePrimaryKeyDDL, PostgresEmitsTableLevelPrimaryKey) {
     EXPECT_EQ(
-            (storm::create_table_sql<OrderItem, storm::orm::schema::Dialect::PostgreSQL>()), order_item_ddl("BIGINT")
+            (storm::create_table_sql<OrderItem, storm::orm::schema::Dialect::PostgreSQL>()), order_item_ddl("INTEGER")
     );
 }
